src/main/kotlin/no/nav/familie/ef/sak/repository/BehandlingRepository.kt
package no.nav.familie.ef.sak.repository

import no.nav.familie.ef.sak.repository.domain.Behandling
import no.nav.familie.ef.sak.repository.domain.BehandlingStatus
import org.springframework.data.jdbc.repository.query.Query
import org.springframework.stereotype.Repository
import java.util.*

@Repository
interface BehandlingRepository : RepositoryInterface<Behandling, UUID>, InsertUpdateRepository<Behandling> {

    fun findByFagsakId(fagsakId: UUID): List<Behandling>

    fun findByFagsakIdAndAktivIsTrue(fagsakId: UUID): Behandling?

    fun findByFagsakIdAndStatus(fagsakId: UUID, status: BehandlingStatus): List<Behandling>

    @Query("SELECT b.*, be.id AS eksternId_id " +
           "        FROM behandling b " +
           "        JOIN behandling_ekstern be on be.behandling=b.id " +
           "        WHERE be.id = :eksternId")
    fun finnMedEksternId(eksternId: Long): Behandling

}
